dl_read_gcp: Data Load, Read file from Google Cloud

Read a single file from Google Cloud Storage (GSC) into a data table

dl_read_gcp(
  path,
  sep = "\t",
  header = TRUE,
  tmpdir,
  gsutil_path = "gsutil",
  check_first = TRUE,
  verbose = FALSE,
  ignore_std_err = TRUE,
  ignore_std_out = TRUE,
  ...
)

Arguments

path

(char) GCS path, i.e., starts with "gs://"

sep

(char) column separator to use with data.table::fread

header

(bool) whether input file has a header line

tmpdir

(char) scratch directory to download files from GCS

gsutil_path

(char) path to gsutil on your computer. Can be "gsutil" if gsutil is in your $PATH.

check_first

(char) check if file exists in tmpdir before downloading it. Read in existing file if it exists. Should be set to TRUE if you are running this function in parallel.

verbose

(logical) TRUE shows messages (default FALSE)

ignore_std_err

(logical) corresponds to ignore.stderr in cmd (default TRUE)

ignore_std_out

(logical) corresponds to ignore.stdout in cmd (default TRUE)

...

optional arguments for data.table::fread

Value

a data table

Details

There is a known issue for Windows users that can occur when dl_read_gcp is run in a local R session where gsutils can not find files in the data hub. Setting either ignore_std_err to FALSE of ignore_std_out to FALSE should fix this.

Examples

if (FALSE) {
pheno = dl_read_gcp(path = "gs://your-bucket/file.txt")
}